Fix: Mantener decimales en consumo de stock cuando hay conversiones de unidades
Cuando se vende un producto fraccionado en una unidad diferente a la de almacenamiento,
la cantidad se convierte (ej: 2 UN / 6 = 0.333... PAQUETE).

El bug estaba en que se hacía casting a int: -(int)0.333... = -0
Por eso no se descont nada del inventario.

Ahora se preservan los decimales: -(float)0.333... = -0.333...
Permitiendo que el inventario se actualice correctamente con valores fraccionarios.

Ejemplo:
- Venta: 2 UNIDADES (unidad_venta_id: 8)
- Conversión: 1 PAQUETE = 6 UN
- Descuento: 2 / 6 = 0.333... PAQUETE
- Stock antes: 22 PAQUETES
- Stock después: 21.666... PAQUETES ✅

+++ app/Services/Venta/VentaDistribucionService.php
@@ -316,7 +316,7 @@ class VentaDistribucionService
                         // ✅ NUEVO: Usar MovimientoStockService que valida y actualiza atomicamente
                         $movimientoRegistrado = $movimientoStockService->registrarMovimientoYActualizar(
                             stockProductoId: $stock->id,
-                            cantidad: -(int)$cantidadTomar,  // Negativo: salida
+                            cantidad: -(float)$cantidadTomar,  // Negativo: salida (mantener decimales para conversiones)
                             tipo: MovimientoInventario::TIPO_SALIDA_VENTA,
                             referencia_tipo: 'venta',
                             referencia_id: $ventaId ?? 0,  // ✅ CORREGIDO (2026-06-29): Usar venta_id real
